# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorAction": {
        "enum": [
          "retry",
          "reload",
          "wait",
          "view_diff",
          "open_resync",
          "unlock_vault",
          "open_provider_settings",
          "edit_instruction",
          "change_model",
          "adjust_budget"
        ],
        "title": "ErrorAction",
        "type": "string"
      },
      "ErrorCode": {
        "enum": [
          "VALIDATION",
          "REVISION_CONFLICT",
          "WORK_BLOCKED",
          "WORK_BUSY_QUEUED",
          "WORK_ACTIVE_JOB",
          "CHAPTER_RANGE_CONFLICT",
          "CHAPTER_READ_ONLY",
          "CANDIDATE_NOT_READY",
          "CANDIDATE_CLOSED",
          "CHAPTER_IS_BASE",
          "STATE_SNAPSHOT_MISSING",
          "CHAPTER_EMPTY",
          "VAULT_LOCKED",
          "VAULT_PASSWORD_INVALID",
          "VAULT_UNLOCK_THROTTLED",
          "VAULT_ALREADY_EXISTS",
          "VAULT_NOT_FOUND",
          "VAULT_CORRUPT",
          "SECRET_MISSING",
          "PROVIDER_AUTH",
          "PROVIDER_UNREACHABLE",
          "PROVIDER_SERVER_ERROR",
          "PROVIDER_RATE_LIMIT",
          "PROVIDER_IN_USE",
          "PROVIDER_REFUSAL",
          "OUTPUT_TRUNCATED",
          "STRUCTURED_OUTPUT_INVALID",
          "BUDGET_EXCEEDED",
          "UNAUTHORIZED",
          "FORBIDDEN_HOST",
          "FORBIDDEN_ORIGIN",
          "NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "DB_BUSY",
          "BACKEND_SHUTTING_DOWN",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorAction"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "StyleProfileOut": {
        "properties": {
          "banned_phrases": {
            "items": {
              "type": "string"
            },
            "title": "Banned Phrases",
            "type": "array"
          },
          "dialogue_dash_char": {
            "anyOf": [
              {
                "enum": [
                  "–",
                  "—"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialogue Dash Char"
          },
          "dialogue_style": {
            "enum": [
              "dash",
              "quotes"
            ],
            "title": "Dialogue Style",
            "type": "string"
          },
          "punctuation_rules": {
            "additionalProperties": true,
            "title": "Punctuation Rules",
            "type": "object"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "tone_mark_style": {
            "enum": [
              "old",
              "new"
            ],
            "title": "Tone Mark Style",
            "type": "string"
          },
          "vocab_register": {
            "enum": [
              "han_viet",
              "balanced",
              "thuan_viet"
            ],
            "title": "Vocab Register",
            "type": "string"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          },
          "voice_samples": {
            "items": {
              "type": "string"
            },
            "title": "Voice Samples",
            "type": "array"
          }
        },
        "required": [
          "vocab_register",
          "dialogue_style",
          "dialogue_dash_char",
          "tone_mark_style",
          "punctuation_rules",
          "banned_phrases",
          "voice",
          "voice_samples",
          "revision"
        ],
        "title": "StyleProfileOut",
        "type": "object"
      },
...
      "WorkCreate": {
        "properties": {
          "genre": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "genre_label_custom": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre Label Custom"
          },
          "language": {
            "const": "vi",
            "default": "vi",
            "title": "Language",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "wizard_step": {
            "anyOf": [
              {
                "enum": [
                  "basics",
                  "brief",
                  "foundation",
                  "address_rules",
                  "event_outline",
                  "writing_config",
                  "review"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "basics",
            "title": "Wizard Step"
          }
        },
        "required": [
          "title"
        ],
        "title": "WorkCreate",
        "type": "object"
      },
...
      "WorkListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "WorkListResponse",
        "type": "object"
      },
      "WorkOut": {
        "properties": {
          "autowrite_mode_default": {
            "enum": [
              "auto",
              "review_each",
              "review_every_k"
            ],
            "title": "Autowrite Mode Default",
            "type": "string"
          },
          "brief": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brief"
          },
          "budget_daily_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Daily Tokens"
          },
          "budget_daily_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Daily Usd"
          },
          "chapter_length_max": {
            "title": "Chapter Length Max",
            "type": "integer"
          },
          "chapter_length_min": {
            "title": "Chapter Length Min",
            "type": "integer"
          },
          "continuity_chapter_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Continuity Chapter No"
          },
          "continuity_reason": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Continuity Reason"
          },
          "continuity_status": {
            "enum": [
              "ok",
              "blocked_needs_resync",
              "stale_from"
            ],
            "title": "Continuity Status",
            "type": "string"
          },
          "cover_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Asset Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "genre_label_custom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre Label Custom"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "const": "vi",
            "title": "Language",
            "type": "string"
          },
          "last_opened_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Opened At"
          },
          "max_repair_rounds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Repair Rounds"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "review_every_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Every K"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "enum": [
              "draft",
              "ready",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "style_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StyleProfileOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_chapters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Chapters"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "wizard_completed_steps": {
            "items": {
              "enum": [
                "basics",
                "brief",
                "foundation",
                "address_rules",
                "event_outline",
                "writing_config",
                "review"
              ],
              "type": "string"
            },
            "title": "Wizard Completed Steps",
            "type": "array"
          },
          "wizard_step": {
            "anyOf": [
              {
                "enum": [
                  "basics",
                  "brief",
                  "foundation",
                  "address_rules",
                  "event_outline",
                  "writing_config",
                  "review"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wizard Step"
          }
        },
        "required": [
          "id",
          "project_id",
          "title",
          "language",
          "genre",
          "genre_label_custom",
          "status",
          "brief",
          "target_chapters",
          "chapter_length_min",
          "chapter_length_max",
          "autowrite_mode_default",
          "review_every_k",
          "max_repair_rounds",
          "budget_daily_usd",
          "budget_daily_tokens",
          "continuity_status",
          "continuity_chapter_no",
          "continuity_reason",
          "cover_asset_id",
          "wizard_step",
          "wizard_completed_steps",
          "last_opened_at",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "WorkOut",
        "type": "object"
      },
      "WorkPatch": {
        "properties": {
          "autowrite_mode_default": {
            "anyOf": [
              {
                "enum": [
                  "auto",
                  "review_each",
                  "review_every_k"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Autowrite Mode Default"
          },
          "brief": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brief"
          },
          "budget_daily_tokens": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Daily Tokens"
          },
          "budget_daily_usd": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Daily Usd"
          },
          "chapter_length_max": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Length Max"
          },
          "chapter_length_min": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Length Min"
          },
          "cover_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Asset Id"
          },
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "genre": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "genre_label_custom": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre Label Custom"
          },
          "max_repair_rounds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Repair Rounds"
          },
          "review_every_k": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Every K"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "draft",
                  "ready",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "target_chapters": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Chapters"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "wizard_completed_steps": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "basics",
                    "brief",
                    "foundation",
                    "address_rules",
                    "event_outline",
                    "writing_config",
                    "review"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wizard Completed Steps"
          },
          "wizard_step": {
            "anyOf": [
              {
                "enum": [
                  "basics",
                  "brief",
                  "foundation",
                  "address_rules",
                  "event_outline",
                  "writing_config",
                  "review"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wizard Step"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "WorkPatch",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WriteStoryApp backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/languages": {
      "get": {
        "operationId": "list_languages",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List Languages",
        "tags": [
          "works"
        ]
      }
    },
...
    "/v1/languages/{code}/genres": {
      "get": {
        "operationId": "list_genres",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List Genres",
        "tags": [
          "works"
        ]
      }
    },
...
    "/v1/works": {
      "get": {
        "operationId": "list_works",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "genre",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Genre"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "ready",
                    "archived"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "continuity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ok",
                    "blocked_needs_resync",
                    "stale_from"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Continuity"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated",
              "enum": [
                "updated",
                "opened",
                "title"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List Works",
        "tags": [
          "works"
        ]
      },
      "post": {
        "operationId": "create_work",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Create Work",
        "tags": [
          "works"
        ]
      }
    },
    "/v1/works/{work_id}": {
      "delete": {
        "operationId": "delete_work",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Delete Work",
        "tags": [
          "works"
        ]
      },
      "get": {
        "operationId": "get_work",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Work",
        "tags": [
          "works"
        ]
      },
      "patch": {
        "operationId": "patch_work",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Patch Work",
        "tags": [
          "works"
        ]
      }
    },
...
    "/v1/works/{work_id}/open": {
      "post": {
        "operationId": "open_work",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Open Work",
        "tags": [
          "works"
        ]
      }
    },
...
    "/v1/works/{work_id}/style-profile": {
      "get": {
        "operationId": "get_style_profile",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Style Profile",
        "tags": [
          "works"
        ]
      },
      "put": {
        "operationId": "put_style_profile",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Put Style Profile",
        "tags": [
          "works"
        ]
      }
    },
...
  }
}